photo-compliance-api-poc/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
opencv-python>=4.8.0.0
Pillow>=10.0.0
numpy>=1.26.0
pydantic>=2.6.0
python-multipart>=0.0.9

mediapipe>=0.10.0

# tests
pytest>=8.0.0
httpx>=0.27.0

